Converted all CLIs to standalone
