+++ server/slicer_cli_list.json
@@ -1,5 +1,5 @@
 {
-  "StandaloneColorDeconvolution": {
+  "ColorDeconvolution": {
     "type"    : "python"
   }
 }
